Changed syntax of import command so that default behavior is to always create a work directory. Added test to prevent accidential nested work directories.

+++ macrotests/test_import.sh
@@ -16,7 +16,7 @@ NOTICE: Nothing was imported.
 EOF
 
 $BOAR import -nq Import TestSession >output.txt 2>&1 || exit 1
-
+test -e Import/.boar && { echo "Dry-run should not create workdir"; exit 1; }
 txtmatch.py expected.txt output.txt || {
     cat output.txt; echo "Dry-run import gave unexpected message"; exit 1; }
 rm output.txt || exit 1
@@ -29,7 +29,7 @@ Checked in session id 2
 EOF
 
 $BOAR import -wq Import TestSession >output.txt 2>&1 || exit 1
-
+test -e Import/.boar || { echo "Normal import should create a workdir"; exit 1; }
 txtmatch.py expected.txt output.txt || {
     echo "Import gave unexpected message"; exit 1; }
 
@@ -43,17 +43,19 @@ txtmatch.py expected.txt output.txt || {
     echo "Unexpected output for workdir status after import -w"; exit 1; }
 
 echo --- Test unchanged import
+rm -r Import/.boar || exit 1
 cat >expected.txt <<EOF
 NOTICE: Nothing was imported.
 !Finished in (.*) seconds
 EOF
 
-$BOAR import -wq Import TestSession >output.txt 2>&1 || exit 1
+$BOAR import -q Import TestSession >output.txt 2>&1 || { cat output.txt; exit 1; }
 
 txtmatch.py expected.txt output.txt || {
     echo "Unchanged import gave unexpected message"; exit 1; }
 
 echo --- Test unchanged import with --allow-empty
+rm -r Import/.boar || exit 1
 cat >expected.txt <<EOF
 Checked in session id 3
 !Finished in (.*) seconds
@@ -75,7 +77,7 @@ EOF
 # Issue 61: AssertionError when importing with an offset ending with a slash
 mkdir MoreImport || exit 1
 echo "More data" >MoreImport/another_file.txt || exit 1
-$BOAR import -q MoreImport TestSession/subpathwithslash/ || 
+$BOAR import -Wq MoreImport TestSession/subpathwithslash/ || 
 { echo "Error when importing to offset path with ending slash"; exit 1; }
 
 
@@ -99,10 +101,33 @@ date >Abc/elephant.txt
 date >Abc/aadvark.txt
 date >Abc/bumblebee.txt
 
-$BOAR import -q Abc TestSession/abc >output.txt 2>&1 || exit 1
+$BOAR import -Wq Abc TestSession/abc >output.txt 2>&1 || exit 1
 
 txtmatch.py expected.txt output.txt || {
     cat output.txt; echo "Abc import gave unexpected message"; exit 1; }
+
+echo --- Test workdir creation options
+
+mkdir NotWorkdir || exit 1
+echo "Avocado" >NotWorkdir/newfile.txt || exit 1
+
+$BOAR import -wWq NotWorkdir TestSession/notworkdir && { echo "Conflicting workdir options should fail"; exit 1; }
+test -e NotWorkdir/.boar && { echo "Conflicting workdir options should not create a workdir"; exit 1; }
+
+$BOAR import -Wq NotWorkdir TestSession/notworkdir1 || { echo "Import with --no-workdir option failed"; exit 1; }
+test -e NotWorkdir/.boar && { echo "Import with --no-workdir should not create a workdir"; exit 1; }
+
+$BOAR import -nq NotWorkdir TestSession/notworkdir1 || { echo "Import with --dry-run option failed"; exit 1; }
+test -e NotWorkdir/.boar && { echo "Import with --dry-run should not create a workdir"; exit 1; }
+
+$BOAR import -wq NotWorkdir TestSession/notworkdir2 || { echo "Import with --create-workdir option failed"; exit 1; }
+test -e NotWorkdir/.boar || { echo "Import with explicit workdir creation should create a workdir"; exit 1; }
+rm -r NotWorkdir/.boar || exit 1
+
+$BOAR import -q NotWorkdir TestSession/notworkdir3 || { echo "Import without options failed"; exit 1; }
+test -e NotWorkdir/.boar || { echo "Import with no options should create a workdir"; exit 1; }
+
+$BOAR import -q NotWorkdir TestSession/notworkdir4 && { echo "Repeated import should fail"; exit 1; }
 
 true
